refactor class components
Co-Authors: @VHAA1718
