+++ components/products/ProductList.tsx
@@ -1,7 +1,7 @@
-import { Grid } from "@mui/material";
 import { FC } from "react"
+import { Grid } from "@mui/material";
 import { IProduct } from "../../interfaces";
-import { ProductCard } from "./ProductCard";
+import { ProductCard } from "./";
 
 interface Props {
     products: IProduct[];
